% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_bbox}
\alias{gh_bbox}
\title{Get bounding boxes}
\usage{
gh_bbox(ghs, crs = 4236)
}
\arguments{
\item{ghs}{(character vector) of geohashes}

\item{crs}{one of (i) character: a string accepted by GDAL, (ii) integer, a
valid EPSG value (numeric), or (iii) an object of class crs}
}
\value{
list of 'bbox'es
}
\description{
Vectorized function to obtain sf bbox class object for a set of
  geohashes
}
\examples{
gh_bbox(c("tdr1w", "tdr1q"))
}
