% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_cover_}
\alias{gh_cover_}
\title{Get a set of geohashes covering a sf object}
\usage{
gh_cover_(x, precision = NULL)
}
\arguments{
\item{x}{(sf object)}

\item{precision}{(integer) precision on the covering geohashes}
}
\value{
(character vector) of covering geohashes
}
\description{
Get a set of geohashes covering a sf object
}
