% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_dist}
\alias{gh_dist}
\title{Compute dist object for a set of geohashes}
\usage{
gh_dist(ghs, ...)
}
\arguments{
\item{ghs}{(character vector) geohash vector}

\item{...}{arguments to be passed to geodist::geodist()}
}
\value{
dist object
}
\description{
set of geohashes is expected to be unique
}
\examples{
gh_dist(c("tdrw1", "tdrw2", "tdrw3"))
}
