% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_distances}
\alias{gh_distances}
\title{Compute distances between a pair of geohash vectors}
\usage{
gh_distances(gh_1, gh_2, ...)
}
\arguments{
\item{gh_1}{(character vector) first geohash vector}

\item{gh_2}{(character vector) second geohash vector}

\item{...}{arguments to be passed to geodist::geodist()}
}
\value{
(numeric vector) matrix of cartesian product of distances with
  dimension of length(gh_1) X length(gh_2)
}
\description{
If one of the vector is a singleton and the other has length >
  1, then the singleton vector is extended to that length of the longer one
}
\examples{
gh_distances(c("tdrw1", "tdrw2"), c("tdrw3", "tdrw4"))
}
