% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_envelope_bbox}
\alias{gh_envelope_bbox}
\title{Single bounding box for a set of geohashes}
\usage{
gh_envelope_bbox(ghs, crs = 4236)
}
\arguments{
\item{ghs}{(character vector) of geohashes}

\item{crs}{one of (i) character: a string accepted by GDAL, (ii) integer, a
valid EPSG value (numeric), or (iii) an object of class crs}
}
\value{
object of class 'bbox'
}
\description{
as sf bbox object
}
\examples{
gh_envelope_bbox(c("tdr1w", "tdr1q"))
mapview::mapview(gh_envelope_bbox(c("tdr1w", "tdr1q")))
}
