% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_expand}
\alias{gh_expand}
\title{Get one level deeper geohashes inside a geohash}
\usage{
gh_expand(ghs)
}
\arguments{
\item{ghs}{(character vector) Input geohashes}
}
\value{
(List) of character vectors containing one level deeper geohashes
}
\description{
Vectorized listing of all geohashes of one level lower
  resolution
}
\examples{
to_test = c("tdxw1", "tdr1u")
gh_expand(to_test)
}
