% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_has_precision}
\alias{gh_has_precision}
\title{Check precision of geohashes}
\usage{
gh_has_precision(ghs, precision)
}
\arguments{
\item{ghs}{(character vector) Input geohashes}

\item{precision}{(integer vector) Either a single precision value or a vector
of length equal to geohashes}
}
\value{
TRUE or FALSE When precision has length 1. Else, a logical vector of
  length equal to number of geohashes
}
\description{
Vectorized checking of precision of geohashes
}
\examples{
gh_has_precision(c("tdxw1", "tdr1"), 5)
#gh_has_precision(c("tdxw1", "tdr1"), precision = c(5, 4))
}
