% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_hull}
\alias{gh_hull}
\title{convex hull of geohashes}
\usage{
gh_hull(ghs)
}
\arguments{
\item{ghs}{(character vector) of geohashes}
}
\value{
Object of class "sfc_POINT"
}
\description{
Obtain sf POINT geometry of convex hull covering the nodes of a set of geohashes
}
\examples{
temp = c("tdr1w", "tdr3qt", "tdr1qw")
hull_sfc = gh_hull(temp)
mapview::mapview(list(hull_sfc, geohashTools::gh_to_sf(temp)))
}
