% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_is_interior}
\alias{gh_is_interior}
\title{check whether gh_2 is in the proper interior of gh_1}
\usage{
gh_is_interior(gh_1, gh_2)
}
\arguments{
\item{gh_1}{first geohash vector}

\item{gh_2}{second geohash vector}
}
\value{
(logical vector) with length equal to longer of the two geohash
  vectors
}
\description{
vectorized check whether gh_2 is in the proper interior of gh_1
}
\examples{
gh_is_interior("tdr1w", c("tdr1wd", "tdr1w1"))
mapview::mapview(geohashTools::gh_to_sf(c("tdr1w", "tdr1wd", "tdr1w1")))
}
