% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_is_interior_}
\alias{gh_is_interior_}
\title{check whether gh_2 is in the proper interior of gh_1}
\usage{
gh_is_interior_(gh_1, gh_2)
}
\arguments{
\item{gh_1}{first geohash}

\item{gh_2}{second geohash}
}
\value{
(flag)
}
\description{
check whether gh_2 is in the proper interior of gh_1
}
