% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_is_neighbor}
\alias{gh_is_neighbor}
\title{check neighborhood relationship between geohash vectors}
\usage{
gh_is_neighbor(gh_1, gh_2)
}
\arguments{
\item{gh_1}{first geohash}

\item{gh_2}{second geohash}
}
\value{
(logical vector) with length equal to longer of the two geohash
  vectors
}
\description{
vectorized over geohash vectors
}
\examples{
gh_is_neighbor(c("tdrw1", "tdrw5"), c("tdrw2","tdrw3"))
mapview::mapview(geohashTools::gh_to_sf(c(c("tdrw1", "tdrw5"), c("tdrw2","tdrw3"))))
}
