% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_is_neighbor_}
\alias{gh_is_neighbor_}
\title{check neighborhood relationship btween a geohash pair}
\usage{
gh_is_neighbor_(gh_1, gh_2)
}
\arguments{
\item{gh_1}{first geohash}

\item{gh_2}{second geohash}
}
\value{
(flag)
}
\description{
check neighborhood relationship btween a geohash pair
}
