% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_is_valid}
\alias{gh_is_valid}
\title{Is a string is a valid geohash}
\usage{
gh_is_valid(ghs)
}
\arguments{
\item{ghs}{(character vector) Input strings}
}
\value{
(Logical vector) with same length as input
}
\description{
Vectorized check if a string is a valid geohash
}
\examples{
to_test = c("tdxw1", "notgeohash")
gh_is_valid(to_test)
}
