% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_pad}
\alias{gh_pad}
\title{Pad geohashes with neighboring one-level smaller geohashes}
\usage{
gh_pad(ghs, self = TRUE)
}
\arguments{
\item{ghs}{(character vector) geohashes}

\item{self}{(flag) Whether the input geohash should be included in the output}
}
\value{
(list of character vectors) of padding geohashes
}
\description{
Pad geohashes with neighboring one-level smaller geohash
}
\examples{
gh_pad(c("tdr1w", "tdr1q"))
}
