% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_pad_}
\alias{gh_pad_}
\title{Pad a geohash with neighboring one-level smaller geohashes}
\usage{
gh_pad_(gh, self = TRUE)
}
\arguments{
\item{gh}{(character vector of length 1) geohash}

\item{self}{(flag) Whether the input geohash should be included in the output}
}
\value{
(character vector) of padding geohashes
}
\description{
Pad a geohash with neighboring one-level smaller geohash
}
