% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_pairwise_distances}
\alias{gh_pairwise_distances}
\title{Compute pairwise distance between a pair of geohash vectors}
\usage{
gh_pairwise_distances(gh_1, gh_2, ...)
}
\arguments{
\item{gh_1}{(character vector) first geohash vector}

\item{gh_2}{(character vector) second geohash vector}

\item{...}{arguments to be passed to geodist::geodist()}
}
\value{
(numeric vector) Vector of pairwise distances with length equal to
  the longest input vector
}
\description{
If one of the vector is a singleton and the other has length >
  1, then the singleton vector is extended to that length of the longer one
}
\examples{
gh_pairwise_distances(c("tdrw1", "tdrw2"), c("tdrw3", "tdrw4"))
}
