% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_integerish}
\alias{is_integerish}
\title{is a number convertible to integer}
\usage{
is_integerish(x)
}
\arguments{
\item{x}{input}
}
\value{
(flag)
}
\description{
is a number convertible to integer
}
\details{
from hadley's assertthat
}
